+++ R/utils__jwt.R
@@ -331,6 +331,7 @@ validate_id_token <- function(
               pins = pins,
               pin_mode = pin_mode,
               min_interval = 30,
+              issuer_match = provider_issuer_match(prov),
               jwks_host_issuer_match = isTRUE(try(
                 prov@jwks_host_issuer_match,
                 silent = TRUE